Split product description block into multiple files
The product description block has been split into three separate files for better organization and maintainability:
- constants.ts: Contains all constants, interfaces and types
- edit.tsx: Contains the block edit component and its wrapper
- index.tsx: Contains block registration and filter setup
